+++ src/modules/usuarios/usuarios.ctrl.js
@@ -1,12 +1,12 @@
 (function (ng) {
     var mod = ng.module('usuariosModule');
 
-    mod.controller('usuariosCtrl', ['usuariosService', '$modal', 'cookiesService', 'dateFilter',
-                                    function (svc, $modal, cookiesSvc, dateFilter) {    	
+    mod.controller('usuariosCtrl', ['usuariosService', '$modal', 'cookiesService', 'dateFilter', 'confirmacionService',
+                                    function (svc, $modal, cookiesSvc, dateFilter, confirmacion) {    	
     	
         this.nuevoUsuario = function() {
             console.log("Entra");
-            this.abrirItem({tipoUsuario: "", nombre: "", fechaNacimiento: "", email: "", password: ""});
+            this.abrirItem({tipoUsuario: "", username:"", first_name: "", last_name:"",  email: "", password: ""});
         }
 
         this.abrirItem = function(usuario) {
@@ -18,10 +18,12 @@
                         usuario: function() {
                             return {
                                 id: usuario.id, 
-                                tipoUsuario: usuario.tipoUsuario, 
-                                nombre: usuario.nombre, 
-                                fechaNacimiento: new Date (usuario.fechaNacimiento), 
-                                email: usuario.email, 
+                                groups: usuario.groups,
+                                tipoUsuario: usuario.tipoUsuario == "" ? usuario.tipoUsuario : (usuario.groups.indexOf(2)!=-1 ? "Administrador" : (usuario.groups.indexOf(1)!=-1 ? "Funcionario" : "Usuario")), 
+                                username: usuario.username,
+                                first_name: usuario.first_name,
+                                last_name: usuario.last_name, 
+                                email: usuario.email,
                                 password: usuario.password
                             };
                         }
@@ -30,7 +32,12 @@
 
             modalInstance.result.then(function(usuario) {
                 console.log(usuario);
-                usuario.fechaNacimiento = dateFilter(usuario.fechaNacimiento, 'yyyy-MM-dd');
+                usuario.groups=[];
+                if(usuario.tipoUsuario=="Administrador") {
+                    usuario.groups.push(2);
+                }else if(usuario.tipoUsuario=="Funcionario"){
+                    usuario.groups.push(1);
+                }
                 if(!usuario.id) {
                     svc.nuevoUsuario(usuario).then(
                         function(data) {
@@ -52,17 +59,26 @@
     	};
 
         this.eliminarItem = function(item) {
-            if(cookiesSvc.getCookieDeAutorizacion().id != item.id) {
-                svc.eliminarUsuario(item).then(function(data) {
-                    this.refrescarUsuarios();
+            confirmacion.showModal({}, this.opcionesEliminacion)
+                .then(function (result) {
+                    if(cookiesSvc.getCookieDeAutorizacion().id != item.id) {
+                        svc.eliminarUsuario(item).then(function(data) {
+                            this.refrescarUsuarios();
+                        }.bind(this));
+                    }
+                    else {
+                        alert("¡POR FAVOR NO SE ELIMINE A USTED MISMO! ¡EL MUNDO SE VA A ACABAR!");
+                    }
                 }.bind(this));
-            }
-            else {
-                alert("¡POR FAVOR NO SE ELIMINE A USTED MISMO! ¡EL MUNDO SE VA A ACABAR!");
-            }
         }
 
         this.usuarios = [];
+        this.opcionesEliminacion = {
+            closeButtonText: 'Cancelar',
+            actionButtonText: 'Eliminar',
+            headerText: 'Confirmar Eliminación',
+            bodyText: '¿Seguro que desea continuar?'
+        };
     	this.refrescarUsuarios();
    }]);
 })(window.angular);
